Rename ebpf_l2_hook.* to ebpf_wfp_ext.*
